move specs to cpp file
